+++ httpclient-tests/src/test/java/io/fabric8/kubernetes/client/http/JettyHttpClientTest.java
@@ -24,4 +24,9 @@ public class JettyHttpClientTest extends OkHttpClientTest {
     return new JettyHttpClientFactory();
   }
 
+  @Override
+  void testConsumeLines() throws Exception {
+    // line parsing not yet supported
+  }
+
 }