+++ junit/kubernetes-server-mock/src/test/java/io/fabric8/kubernetes/client/server/mock/crud/KubernetesCrudDispatcherPutTest.java
@@ -113,7 +113,8 @@ class KubernetesCrudDispatcherPutTest {
     final HttpResponse<String> result = httpClient.sendAsync(request, String.class).get(10, TimeUnit.SECONDS);
     // Then
     assertThat(result)
-        .hasFieldOrPropertyWithValue("response.code", 400)
+        .extracting(HttpResponse::code).isEqualTo(400);
+    assertThat(result)
         .extracting(HttpResponse::body).asString()
         .contains("the name of the object (different) does not match the name on the URL (mismatched-name)");
   }